+++ .gitignore
@@ -2,4 +2,5 @@ src\creds.json
 src/creds.json
 update_times.json
 logs
+database
 staff_bulletin.txt
