--- database/2024-01-04-wager_database.json
@@ -1,23 +1,0 @@
-{
-    "wagerID": 2,
-    "1803768": {
-        "id": "1803768",
-        "type": "BET",
-        "customer_ref": "DRADS007",
-        "details": {
-            "selections": [
-                [
-                    "Wolverhampton, 19:30 - Betuk. Home Of T.-Fenwa Handica, Miss Bluebelle",
-                    8.0
-                ]
-            ],
-            "timestamp": "19:11:18",
-            "risk_category": "W",
-            "bet_details": "Win",
-            "unit_stake": "\u00a350.00",
-            "payment": "\u00a350.00",
-            "bet_type": "1 Singles"
-        },
-        "wagerID": 1
-    }
-}
